+++ app/src/main/AndroidManifest.xml
@@ -11,11 +11,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-
-        <activity android:name=".LeaderBoardActivity"></activity>
-        <activity android:name=".GameActivity" ></activity>
-        <activity android:name=".WebViewActivity"></activity>
-        <activity android:name=".MainActivity"></activity>
+        <activity android:name=".LeaderBoardActivity" />
+        <activity android:name=".GameActivity" />
+        <activity android:name=".WebViewActivity" />
+        <activity android:name=".MainActivity" />
         <activity android:name=".SplashScreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -23,6 +22,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
