+++ app/src/main/res/layout/activity_leader_board.xml
@@ -34,7 +34,7 @@
             android:textAlignment="center"
             android:layout_weight="1"/>
         <TextView
-            android:id="@+id/scoreboardName"
+            android:id="@+id/leaderBoardName"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:textSize="22sp"
@@ -44,7 +44,7 @@
             android:textAlignment="center"
             android:layout_weight="1"/>
         <TextView
-            android:id="@+id/scoreboardPoints"
+            android:id="@+id/leaderBoardScore"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:textSize="22sp"
