+++ app/src/main/res/values/strings.xml
@@ -17,6 +17,6 @@
     <string name="leader_board_title">Leader Board</string>
     <string name="leader_board_table_header_rank">Rank</string>
     <string name="leader_board_table_header_name">Name</string>
-    <string name="leader_board_table_header_score">Time</string>
+    <string name="leader_board_table_header_score">Score</string>
     <string name="leader_board_back_btn">Back</string>
 </resources>
